+++ resources/views/pages/show.blade.php
@@ -1,4 +1,3 @@
 @extends('layouts.public', [
     'title' => $page->title,
-    'metaDescription' => $metaDescription,
 ])